requirements.txt
# 🏗️ Core Flask Framework
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
click==8.1.8
itsdangerous==2.2.0
MarkupSafe==3.0.2
blinker==1.9.0

# 🔐 Authentication & Forms
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1

# 🗄️ Database & ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.38
greenlet==3.1.1

# ⚙️ Environment & Utilities
python-dotenv==1.0.1
typing_extensions==4.12.2
